+++ Code/GraphMol/MolDraw2D/catch_tests.cpp
@@ -4706,48 +4706,3 @@ TEST_CASE("GitHub #5383: cairo error when using similarity maps", "") {
   }
 #endif
 }
-
-TEST_CASE("Force use of MolBlock wedges", "") {
-  auto m = R"CTAB(bad wedging
-  ChemDraw07092209022D
-
-  0  0  0     0  0              0 V3000
-M  V30 BEGIN CTAB
-M  V30 COUNTS 7 7 0 0 0
-M  V30 BEGIN ATOM
-M  V30 1 C -0.714471 0.825000 0.000000 0
-M  V30 2 C -0.714471 0.000000 0.000000 0
-M  V30 3 C -0.000000 -0.412500 0.000000 0
-M  V30 4 C 0.714471 0.000000 0.000000 0
-M  V30 5 C 0.714471 0.825000 0.000000 0
-M  V30 6 C -0.000000 1.237500 0.000000 0
-M  V30 7 C -0.000000 -1.237500 0.000000 0
-M  V30 END ATOM
-M  V30 BEGIN BOND
-M  V30 1 1 1 2
-M  V30 2 1 2 3
-M  V30 3 1 3 4 CFG=1
-M  V30 4 1 4 5
-M  V30 5 1 5 6
-M  V30 6 1 6 1
-M  V30 7 1 3 7
-M  V30 END BOND
-M  V30 END CTAB
-M  END
-)CTAB"_ctab;
-  REQUIRE(m);
-  MolDraw2DSVG drawer(400, 400, -1, -1, NO_FREETYPE);
-  drawer.drawOptions().useMolBlockWedging = true;
-  drawer.drawOptions().addAtomIndices = true;
-  drawer.drawMolecule(*m);
-  drawer.finishDrawing();
-  auto text = drawer.getDrawingText();
-  std::regex regex("class='bond-2 atom-2 atom-3' .*M .*L .*L .*Z");
-  std::smatch triangleMatch;
-  REQUIRE(std::regex_search(text, triangleMatch, regex));
-  REQUIRE(triangleMatch.size() == 1);
-  std::ofstream outs("test_molblock_wedges.svg");
-  outs << text;
-  outs.flush();
-  check_file_hash("test_molblock_wedges.svg");
-}
